renamed console input and output to eval input and output
